Organize the package structure to ease imports
