+++ flaskblog/__init__.py
@@ -1,0 +1,11 @@
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+
+
+
+app = Flask(__name__)
+app.config['SECRET_KEY'] = 'c4ad0f02dad6d26e22ee3ed0a70777cf'
+app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///site.db'
+db = SQLAlchemy(app)
+
+from flaskblog import routes
